requirements.txt
python-docx>=1.0
openpyxl>=3.1
jinja2>=3.1
